# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API de Gestão Escolar",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/teachers/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "Filtro de busca",
            "in": "query",
            "required": false,
            "type": "object",
            "schema": {
              "$ref": "#/components/schemas/teacher"
            },
            "style": "form",
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/teacher"
              }
            }
          }
        }
      }
    },
    "/teachers/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "schoolDiscipline": {
                    "example": "any"
                  },
                  "contact": {
                    "example": "any"
                  },
                  "phoneNumber": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "Filtro de busca",
            "in": "query",
            "required": false,
            "type": "object",
            "schema": {
              "$ref": "#/components/schemas/user"
            },
            "style": "form",
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/user"
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "user": {
                    "example": "any"
                  },
                  "pwd": {
                    "example": "any"
                  },
                  "level": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "user": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Andre Faria Ruaro"
          },
          "email": {
            "type": "string",
            "example": "andre.ruaro@unesc.net"
          },
          "user": {
            "type": "string",
            "example": "andre.ruaro"
          },
          "pwd": {
            "type": "string",
            "example": "7a6cc1282c5f6ec0235acd2bfa780145aaskem5n"
          },
          "level": {
            "type": "string",
            "example": "admin"
          },
          "status": {
            "type": "string",
            "example": "on"
          }
        },
        "required": [
          "name",
          "email",
          "user",
          "pwd"
        ]
      },
...
      "teacher": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Judite Heeler"
          },
          "schoolDiscipline": {
            "type": "string",
            "example": "Artes, Português"
          },
          "contact": {
            "type": "string",
            "example": "j.heeler@gmail"
          },
          "phoneNumber": {
            "type": "string",
            "example": "48 9696 5858"
          },
          "status": {
            "type": "string",
            "example": "on"
          }
        },
        "required": [
          "name",
          "schoolDiscipline",
          "contact",
          "phoneNumber"
        ]
      },
...
    }
  }
}